Refactored renders into seperate files
