+++ lib/generator.js
@@ -3,13 +3,13 @@ const fs = require('fs');
 const xfs = require('fs.extra');
 const minimatch = require('minimatch');
 const parser = require('@asyncapi/parser');
+const {configureReact, renderReact} = require('./renderer/react');
+const {configureNunjucks, renderNunjucks} = require('./renderer/nunjucks');
 const { parse, AsyncAPIDocument } = parser;
 const ramlDtParser = require('@asyncapi/raml-dt-schema-parser');
 const openapiSchemaParser = require('@asyncapi/openapi-schema-parser');
 const avroSchemaParser = require('@asyncapi/avro-schema-parser');
-const Nunjucks = require('nunjucks');
 const jmespath = require('jmespath');
-const AsyncReactSDK = require('asyncAPISDK');
 const filenamify = require('filenamify');
 const git = require('simple-git/promise');
 const npmi = require('npmi');
@@ -197,9 +197,9 @@ class Generator {
    */
   async configureTemplate() {
     if (isReactTemplate(this.templateConfig)) {
-      await this.configureReact();
+      await configureReact(this.templateDir, this.templateContentDir, TRANSPILED_TEMPLATE_LOCATION);
     } else {
-      this.configureNunjucks();
+      this.nunjucks = configureNunjucks(this.debug, this.templateDir);
     }
   }
 
@@ -673,46 +673,6 @@ class Generator {
     }
   }
 
-  /**
-   * Renders the template with nunjucks and returns a string.
-   *
-   * @private
-   * @param {AsyncAPIDocument} asyncapiDocument AsyncAPI document to pass to the template.
-   * @param {String} filePath Path to the file being rendered.
-   * @param {Object} [extraTemplateData] Extra data to pass to the template.
-   * @return {Promise}
-   */
-  renderNunjucks(asyncapiDocument, tempalteString, filePath, extraTemplateData = {}) {
-    return new Promise((resolve, reject) => {
-      this.nunjucks.renderString(tempalteString, {
-        asyncapi: asyncapiDocument,
-        params: this.templateParams,
-        originalAsyncAPI: this.originalAsyncAPI,
-        ...extraTemplateData
-      }, { path: filePath }, (err, result) => {
-        if (err) return reject(err);
-        resolve(result);
-      });
-    });
-  }
-
-  /**
-   * Renders the template with react and returns the  
-   * @private
-   * @param {AsyncAPIDocument} asyncapiDocument AsyncAPI document to pass to the template.
-   * @param {String} filePath Path to the file being rendered.
-   * @param {Object} [extraTemplateData] Extra data to pass to the template.
-   * @return {Promise<import('asyncAPISDK').TemplateRenderResult>}
-   */
-  async renderReact(asyncapiDocument, filePath, extraTemplateData = {}) {
-    filePath = filePath.replace(this.templateContentDir, path.resolve(this.templateDir, TRANSPILED_TEMPLATE_LOCATION));
-    return await AsyncReactSDK.renderTemplate(filePath, {
-      asyncapi: asyncapiDocument,
-      params: this.templateParams,
-      originalAsyncAPI: this.originalAsyncAPI,
-      ...extraTemplateData
-    }, this.debug);
-  }
 
   /**
    * Renders the content of a file and outputs it.
@@ -725,10 +685,10 @@ class Generator {
    */
   async renderFile(asyncapiDocument, filePath, extraTemplateData = {}) {
     if (isReactTemplate(this.templateConfig)) {
-      return await this.renderReact(asyncapiDocument, filePath, extraTemplateData);
+      return await renderReact(asyncapiDocument, filePath, extraTemplateData, this.templateDir, this.templateContentDir, TRANSPILED_TEMPLATE_LOCATION, this.templateParams, this.debug, this.originalAsyncAPI);
     }
     const tempalteString = await readFile(filePath, 'utf8');
-    return this.renderNunjucks(asyncapiDocument, tempalteString, filePath, extraTemplateData);
+    return renderNunjucks(asyncapiDocument, tempalteString, filePath, extraTemplateData, this.templateParams, this.originalAsyncAPI, this.nunjucks);
   }
 
   /**
@@ -757,28 +717,6 @@ class Generator {
     if (!fileExists) return true;
 
     return !this.noOverwriteGlobs.some(globExp => minimatch(filePath, globExp));
-  }
-
-  /**
-   * Configures React templating system, this handles all the transpilation work.
-   * @private
-   */
-  async configureReact() {
-    const outputDir = path.resolve(this.templateDir, `./${TRANSPILED_TEMPLATE_LOCATION}`);
-    await AsyncReactSDK.transpileFiles(this.templateContentDir, outputDir, {
-      recursive: true
-    });
-  }
-
-  /**
-   * Configures Nunjucks templating system
-   * @private
-   */
-  configureNunjucks() {
-    const config = {};
-    if (this.debug) config.dev = true;
-
-    this.nunjucks = new Nunjucks.Environment(new Nunjucks.FileSystemLoader(this.templateDir), config);
   }
 
   /**
